+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
-- `sandstone/Scroller` focus scroll thumb initially when it is used in Panels
+- `sandstone/Scroller` to focus scroll thumb initially when it is used in Panels
 
 ## [2.5.0-alpha.2] - 2022-05-09